backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
boto3==1.34.84
cloudshare-lib==0.1.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic[email]==2.7.1
